seperate taiko host functions for std/no-std
