+++ .vscode/settings.json
@@ -8,6 +8,7 @@
         "./host/Cargo.toml",
         "./testing/ef-tests/testguest/Cargo.toml",
         "./lib/Cargo.toml",
+        "./lib/Cargo.toml",
         "./lib/Cargo.toml"
     ]
 }
