+++ lib/Cargo.toml
@@ -41,7 +41,7 @@ bincode = "1.3"
 serde_with = "3.1"
 
 [features]
-default = ["std", "taiko"]
+default = ["taiko"]
 std = []
 taiko = [
     "zeth-primitives/taiko",
